pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notebook4m"
version = "0.1.0"
requires-python = ">=3.8"
description = "ローカル GGUF モデルで対話するツール"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{ name = "あなたの名前", email = "you@example.com" }]
dependencies = [
  "llama-cpp-python>=0.3.2",
  "prompt_toolkit>=3.0.51",
  "repo-to-text>=0.6.0",
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.optional-dependencies]
dev = ["pytest", "black", "isort"]

[project.scripts]
notebook4m = "notebook4m:main"
